Stalfos spawns with hitbox
