+++ MonoZelda/Enemies/EnemyList.cs
@@ -8,5 +8,16 @@ namespace MonoZelda.Enemies
 {
     public enum EnemyList
     {
+        Aquamentus,
+        Dodongo,
+        Gel,
+        Goriya,
+        Keese,
+        Oldman,
+        Rope,
+        Stalfos,
+        Trap,
+        Wallmaster,
+        Zol
     }
 }
